User_Service/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.12.2
